+++ service-api/service-csharp/identity/src/Identity.Application/DependencyInjection.cs
@@ -17,6 +17,7 @@ public static class DependencyInjection
     services.AddScoped<RemoveBootstrapTeamMember>();
     services.AddScoped<RevokeBootstrapUserRole>();
     services.AddScoped<GetBootstrapCompanyByPublicId>();
+    services.AddScoped<GetBootstrapUserByPublicId>();
     services.AddScoped<UpdateBootstrapTeam>();
     services.AddScoped<UpdateBootstrapCompany>();
     services.AddScoped<UpdateBootstrapUser>();